pyproject.toml
[project]
name = "nezu-works-bot"
version = "0.1.0"
description = "Nezu Works Bot for line-works-sdk."
requires-python = ">=3.13.1"
dependencies = [
    "line-works-sdk",
    "python-dotenv",
    "rich",
    "ruff",
]

[tool.ruff]
target-version = "py313"
line-length = 79

select = [
    "E",   # pycodestyle エラー
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "C90", # mccabe (循環的複雑度)
    "S",   # flake8-bandit (セキュリティ)
    "RET", # flake8-return (return文の一貫性)
    "SIM", # flake8-simplify (コード簡素化)
    "T20", # flake8-print (print文検出)
]

# インデント設定
indent-width = 4

# 除外するファイル/ディレクトリ
exclude = [
    ".git",
    ".ruff_cache",
    "__pycache__",
    "venv",
    ".env",
    ".venv",
]

[tool.ruff.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["app"]
known-third-party = ["flask", "psutil", "requests"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.pydocstyle]
convention = "google"
ignore-decorators = ["overrides"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
ignore = [
    "E402",  # module level import not at top of file
    "S201",  # Flask debug=True security issue
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.mccabe]
max-complexity = 10  # 関数の最大複雑度

# 命名規則の設定
[tool.ruff.pep8-naming]
classmethod-decorators = ["classmethod"]
staticmethod-decorators = ["staticmethod"]

# 行の長さ制限
[tool.ruff.pycodestyle]
max-line-length = 79
max-doc-length = 72

# メトリクス警告
[tool.ruff.pylint]
max-args = 5          # 関数の最大引数数
max-statements = 50   # 関数内の最大文数
max-locals = 15       # 関数内のローカル変数の最大数
max-returns = 6       # return文の最大数
max-branches = 12     # 分岐の最大数
